bro-script: Add simple baseline implementation
This implementation only logs the (combined) request and response events
that occur within the same connection.  This assumes that a response is
always send over the same connection as a request.  It is unclear
whether this assumption really holds.

This implementation does not yet contain error handling, so if there was
no response for a request, Bro displays an error message.

It also does not contain an interpretation of the values, so if multiple
values are read within one request, they are displayed in the same log
entry.